Use extract_numbers for sort keys to preserve previous DB output.
